--- a/client/package.json
+++ b/client/package.json
@@ -179,6 +179,7 @@
     "eslint-plugin-vuejs-accessibility": "^2.2.0",
     "expose-loader": "^4.1.0",
     "fake-indexeddb": "^6.0.0",
+    "fork-ts-checker-webpack-plugin": "^9.0.2",
     "fs-extra": "^11.2.0",
     "gulp": "^4.0.2",
     "ignore-loader": "^0.1.2",