--- a/client/package.json
+++ b/client/package.json
@@ -22,7 +22,7 @@
   "resolutions": {
     "chokidar": "3.5.3",
     "prismjs": "1.29.0",
-    "vue": "2.7.14",
+    "vue": "2.7.16",
     "vue-loader": "15.11.1"
   },
   "dependencies": {
@@ -106,7 +106,7 @@
     "vega": "^5.30.0",
     "vega-embed": "^6.26.0",
     "vega-lite": "^5.21.0",
-    "vue": "^2.7.14",
+    "vue": "^2.7.16",
     "vue-echarts": "^7.0.3",
     "vue-infinite-scroll": "^2.0.2",
     "vue-multiselect": "^2.1.7",
@@ -206,7 +206,7 @@
     "tsconfig-paths-webpack-plugin": "^4.1.0",
     "typescript": "^5.7.3",
     "vue-loader": "^15.11.1",
-    "vue-template-compiler": "^2.7.14",
+    "vue-template-compiler": "^2.7.16",
     "webpack": "^5.98.0",
     "webpack-cli": "^6.0.1",
     "webpack-dev-server": "^5.2.0",