--- a/client/package.json
+++ b/client/package.json
@@ -182,7 +182,7 @@
     "fs-extra": "^11.2.0",
     "gulp": "^4.0.2",
     "ignore-loader": "^0.1.2",
-    "imports-loader": "^4.0.1",
+    "imports-loader": "^5.0.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "jest-fixed-jsdom": "^0.0.2",
@@ -209,12 +209,12 @@
     "vue-template-compiler": "^2.7.14",
     "webpack": "^5.98.0",
     "webpack-cli": "^6.0.1",
-    "webpack-dev-server": "^4.15.1",
-    "webpack-merge": "^5.10.0",
+    "webpack-dev-server": "^5.2.0",
+    "webpack-merge": "^6.0.1",
     "xml-js": "^1.6.11",
     "xml2js": "^0.6.2",
     "yaml-jest": "^1.2.0",
-    "yaml-loader": "^0.8.0"
+    "yaml-loader": "^0.8.1"
   },
   "peerDependencies": {
     "postcss": "^8.4.6"