--- a/client/package.json
+++ b/client/package.json
@@ -146,7 +146,6 @@
     "@babel/plugin-transform-runtime": "^7.23.2",
     "@babel/preset-env": "^7.23.2",
     "@babel/preset-typescript": "^7.23.2",
-    "@cerner/duplicate-package-checker-webpack-plugin": "^2.3.0",
     "@pinia/testing": "0.1.3",
     "@testing-library/jest-dom": "^6.4.8",
     "@types/d3": "^7.4.2",