--- a/client/src/layout/window-manager.js
+++ b/client/src/layout/window-manager.js
@@ -1,5 +1,5 @@
 /** Adds window manager masthead icon and functionality **/
-import "winbox/dist/css/winbox.min.css";
+import "winbox/src/css/winbox.css";
 
 import _l from "utils/localization";
 import { withPrefix } from "utils/redirect";