--- a/client/webpack.config.js
+++ b/client/webpack.config.js
@@ -186,7 +186,6 @@ module.exports = (env = {}, argv = {}) => {
                         },
                         {
                             loader: "css-loader",
-                            options: { sourceMap: true },
                         },
                         {
                             loader: "postcss-loader",