--- a/client/webpack.config.js
+++ b/client/webpack.config.js
@@ -244,6 +244,12 @@ module.exports = (env = {}, argv = {}) => {
                 }),
             }),
         ],
+        cache: {
+            type: "filesystem",
+            buildDependencies: {
+                config: [__filename],
+            },
+        },
         devServer: {
             client: {
                 overlay: {