--- a/client/webpack.config.js
+++ b/client/webpack.config.js
@@ -286,7 +286,7 @@ module.exports = (env = {}, argv = {}) => {
     };
 
     if (process.env.GXY_BUILD_SOURCEMAPS) {
-        buildconfig.devtool = "source-map";
+        buildconfig.devtool = "eval-cheap-module-source-map";
     }
 
     return buildconfig;