Looking for some low hanging fruit to address speed issues folks have mentioned.  

This isn't the long term fix -- now that we've addressed more of the issues, I'll also update the vite branch and see what blockers are still there.

On a clean build in a slower VM, I'm seeing:

```
 SMP  ⏱
General output time took 6 mins, 14.19 secs

 SMP  ⏱  Plugins
DuplicatePackageCheckerPlugin took 0.276 secs
MiniCssExtractPlugin took 0.036 secs
DefinePlugin took 0.002 secs
ProvidePlugin took 0.001 secs
VueLoaderPlugin took 0 secs
DumpMetaPlugin took 0 secs

 SMP  ⏱  Loaders
vue-loader, and
ts-loader, and
vue-loader took 5 mins, 33.4 secs
  module count = 882
vue-loader took 5 mins, 13.33 secs
  module count = 2448
modules with no loaders took 4 mins, 14.085 secs
  module count = 6466
ts-loader took 2 mins, 9.23 secs
  module count = 255
expose-loader took 4.6 secs
  module count = 3
yaml-loader took 2.55 secs
  module count = 4
imports-loader took 0.963 secs
  module count = 12
amdi18n-loader took 0.109 secs
  module count = 1
vue-loader, and
mini-css-extract-plugin, and
css-loader, and
postcss-loader, and
sass-loader, and
vue-loader took 0.094 secs
  module count = 474
mini-css-extract-plugin, and
css-loader, and
postcss-loader, and
sass-loader took 0.006 secs
  module count = 32
  ```
  
  As a starting point.


Just adding the caching gets a rebuild (with dev-server as well) down to this:
```
webpack 5.94.0 compiled with 15 warnings in 11906 ms
```


## How to test the changes?
(Select all options that apply)
- [ ] I've included appropriate [automated tests](https://docs.galaxyproject.org/en/latest/dev/writing_tests.html).
- [ ] This is a refactoring of components with existing test coverage.
- [ ] Instructions for manual testing are as follows:
  1. [add testing steps and prerequisites here if you didn't write automated tests covering all your changes]

## License
- [x] I agree to license these and all my past contributions to the core galaxy codebase under the [MIT license](https://opensource.org/licenses/MIT).