@mvdbeek 2025-03-19
https://github.com/galaxyproject/galaxy/actions/runs/13955102548/job/39064104325?pr=19851 went down to 6 minutes ... did we expect that ? coincidence ?