@dannon 2025-03-19
Yep -- that's from the typescript transpileOnly tweak.  I'm pushing typeChecking into a separate process, I think we can get this pretty fast.  With full cache a rebuild is like 5s locally now.