@jmchilton 2025-03-20
The client linting error count going up is likely due to updated dependencies. I would recommend we merge with that being red and make this the new baseline. We can break up those issues and address them in subsequent PRs and other folks' PR won't be red because this will be the new baseline.