@jmchilton 2025-03-20
> [Circular dependency detector. I regret looking.](https://github.com/galaxyproject/galaxy/pull/19851/commits/9b631663c73e92eac1f5df79c170b9bad313b195)

😄😄😄 . It is a fun list of bugs 🤷‍♀️. Can we separate that into a different change - I really want the speed! You know I love fixing circular dependencies and I'm happy to help with that  - I think https://github.com/galaxyproject/galaxy/pull/19857 will help but I don't think it needs to happen in the context of this PR ideally.