@dannon 2025-03-20
@jmchilton Thanks for taking a look!  I agree w/ the vue-tsc error count diagnosis and suggestion.  

I thought the circular dependencies might be causing additional slowness in the build, which is why I pulled on that thread a bit.  Before merging, I can make circular dependency detector not actually *fail* the build unless the number detected increases similar to vue-tsc errors.