@jmchilton 2025-03-20
> I can make circular dependency detector not actually fail the build unless the number detected increases similar to vue-tsc errors.

Looking through the errors - a lot of them are effectively duplicates. Something imports utils.ts and now you've got a good chance of a circular dependency on very different parts of the app. It isn't nearly as clean in terms of isolation as the typescript errors and would almost discourage best practices like code reuse to keep the number down. I think the same logic means a lot of people would have their PRs have to solve problems that are not thiers. I would suggest we tackle some of main duplicates around foundational things (data.js, userStore, etc..) before imposing that restriction. Maybe I'm wrong though - I haven't applied it to my branches and looked at the results.