@jmchilton 2025-03-20
Latest merged dev branch PR:

> webpack 5.94.0 compiled with 25 warnings in 542172 ms

This PR:

> webpack 5.98.0 compiled with 3 warnings in 67821 ms

A million thanks - this is going to mean a tremendous improvement to my Galaxy coding QOL and productivity.